% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_functions.R
\name{add_border}
\alias{add_border}
\title{Make an enhanced field layout plot with border rows}
\usage{
add_border(
  x,
  row_id = "ROW",
  col_id = "COLUMN",
  rep_id = "REP",
  trt_id = "TREATMENT",
  title = "Field layout",
  text_sz = 3,
  axis_title_sz = 12,
  xlab = "Column",
  ylab = "Row"
)
}
\arguments{
\item{x}{is the input data frame of fieldbook that has row and column 
coordinates of each plot.}

\item{row_id}{is the string column identifier for ROW in the input fieldbook.}

\item{col_id}{is the string column identifier for COLUMN in the input fieldbook.}

\item{rep_id}{is the string column identifier for REP in the input fieldbook.}

\item{trt_id}{is the string column identifier for TREATMENT in the input fieldbook.}

\item{title}{is the title of the field layout plot.}

\item{text_sz}{is the text size to print treatment names on the tiles.}

\item{axis_title_sz}{is the text size for axis titles.}

\item{xlab}{is a string to label x axis; default is 'Column'.}

\item{ylab}{is a string to label y axis; default is 'Row'.}
}
\value{
a field layout with border rows around the entire perimeter.
}
\description{
A helper function that adds border rows to the entire perimeter of a
field laid out in a rectangular or square grid. Each experimental plot must
have a coordinate that is specified by row and column numbers in the 
grid layout.
}
\note{
this function works best with input fieldbooks generated with the FieldHub package
}
\examples{


# Plot a field layout with border rows

library(qrlabelr)

set.seed(123)

add_border(x = data.frame(LOCATION = rep("BAMBEY", 12), 
                         PLOT = c(1001:1012),
                         ROW = c(rep(1, 6), rep(2, 6)),
                         COLUMN = c(1:6, 1:6),
                         REP = rep(1, 12),
                         TREATMENT = sample(paste0("G-", 1:12))),
text_sz = 3)



}
