% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_functions.R
\name{make_qrcode}
\alias{make_qrcode}
\title{Make QR codes}
\usage{
make_qrcode(my_id, ec_level = 3)
}
\arguments{
\item{my_id}{unique ID string to be encoded to QR code}

\item{ec_level}{error correction level (`0` - `3`, lowest to highest)}
}
\value{
a raster grob image of QR code.
}
\description{
A helper function for QR code generation using the qrencoder library.
The function converts the generated QR code into a raster grob image that can
be plotted using the grid.draw function in the grid package.
}
\examples{

\dontrun{
make_qrcode("PLT101-NURSERY", 4)
}


}
