% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbd.R
\docType{data}
\name{rcbd}
\alias{rcbd}
\title{rcbd}
\format{
A data frame with 144 observations and 7 variables:
\describe{
  \item{\code{SN}}{\code{double} contains observation serial numbers.}
  \item{\code{LOCATION}}{\code{character} contains information on plot location.}
  \item{\code{PLOT}}{\code{double} contains the plot number.}
  \item{\code{ROW}}{\code{double} contains the row number.}
  \item{\code{COLUMN}}{\code{double} contains the column number.}
  \item{\code{REP}}{\code{double} contains the replication number.}
  \item{\code{TREATMENT}}{\code{character} contains the treatment identifier.}
}
}
\usage{
data(rcbd)
}
\description{
A sample field book generated with the 'FielDHub' package based on a
Randomized Complete Block Design layout at two locations.
}
\keyword{datasets}
