% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch 'EasyQrlabelr'.}
\usage{
run_app()
}
\value{
No return value, called for side effects. Opens the 'EasyQrlabelr' 'shiny' app in your default browser.
}
\description{
An interface function to launch the 'EasyQrlabelr' 'shiny' app.
}
\examples{
library(qrlabelr)
if(interactive())
qrlabelr::run_app()

}
\concept{"plot labels"}
