% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_lattice.R
\docType{data}
\name{square_lattice}
\alias{square_lattice}
\title{Square Lattice}
\format{
A data frame with 216 observations and 11 variables:
\describe{
  \item{\code{SN}}{\code{double} contains observation serial numbers.}
  \item{\code{LOCATION}}{\code{character} contains information on plot location.}
  \item{\code{PLOT}}{\code{double} contains the plot number.}
  \item{\code{ROW}}{\code{double} contains the row number.}
  \item{\code{COLUMN}}{\code{double} contains the column number.}
  \item{\code{REP}}{\code{double} contains the replication number.}
  \item{\code{IBLOCK}}{\code{double} contains the incomplete block number.}
  \item{\code{UNIT}}{\code{double} contains the unit number.}
  \item{\code{ENTRY}}{\code{double} contains the entry number.}
  \item{\code{TREATMENT}}{\code{character} contains the treatment identifier.}
  \item{\code{SEED_SOURCE}}{\code{character} contains the seed source identifier.}
}
}
\usage{
data(square_lattice)
}
\description{
A sample data set to create field plot and general_purpose labels.
}
\keyword{datasets}
