% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_functions.R
\name{field_label}
\alias{field_label}
\title{Create field plot label function embossed with QR codes}
\usage{
field_label(
  dat,
  get_unique_id = c("ruid", "uuid", "custom"),
  unique_id = NULL,
  Year = NULL,
  rname = NULL,
  Trial = "PYT",
  seed_source = FALSE,
  IBlock = FALSE,
  rep_id = "REP",
  plot_id = "PLOT",
  row_id = "ROW",
  col_id = "COLUMN",
  loc_id = "LOCATION",
  entry_id = "TREATMENT",
  IBlock_id = "IBLOCK",
  seed_source_id = NULL,
  ...
)
}
\arguments{
\item{dat}{is an input data frame of fieldbook that contains plot attributes.
To design field plot labels, the imported fieldbook must have LOCATION, PLOT,
ROW, COLUMN/RANGE, REP, TREATMENT columns. The order of the columns in the fieldbook
is not important, and the columns can be any name the user desires.}

\item{get_unique_id}{Set to 'ruid' if reproducible and informative unique ids
are to be generated from imported fieldbook. Set to 'uuid' if universal unique ids
are to be generated from imported fieldbook. Set to 'custom' if imported fieldbook
already has unique IDs for each plot.}

\item{unique_id}{is the column identifier for UNIQUE_ID in the imported field book}

\item{Year}{is the year of experiment or trial.}

\item{rname}{is the researcher's name. Initials or intials of first and middle
names and the last name.}

\item{Trial}{is the name of the trial to use.}

\item{seed_source}{set to TRUE if seed source is included in the imported fieldbook,
FALSE if otherwise.}

\item{IBlock}{set to TRUE if \code{dat} contains incomplete blocks within REPs.}

\item{rep_id}{is the column identifier for REP in the imported field book.}

\item{plot_id}{is the column identifier for PLOT in the imported field book.}

\item{row_id}{is the column identifier for ROW in the imported field book.}

\item{col_id}{is the column identifier for COLUMN in the imported field book.}

\item{loc_id}{is the column identifier for LOCATION in the imported field book.}

\item{entry_id}{is the column identifier for ENTRY/TREATMENT in the imported field book.}

\item{IBlock_id}{is the column identifier for IBLOCK in the imported field book.
It must be provided if \code{IBlock} is set to TRUE.}

\item{seed_source_id}{is the column identifier for SEED_SOURCE in the imported field book.
It must be provided if \code{seed_source} is set to TRUE.}

\item{...}{additional arguments passed to the \code{create_label()} function.}
}
\value{
A PDF file containing field plot labels affixed with QR codes, and
 an updated fieldbook-- all saved to the default or set working directory.
}
\description{
Create machine and human-readable plot labels that are well-suited for field experiments.
}
\details{
The default column identifiers for LOCATION, PLOT, ROW, COLUMN/RANGE, REP, TREATMENT
are based on the column IDs of a fieldbook generated using the FieldHub package.
If user imports any fieldbook generated with other programs,
the user must specify the equivalent column identifiers used for LOCATION, PLOT, REP,
ROW, COLUMN, and TREATMENT/ENTRY in the imported fieldbook.

if \code{get_unique_id = 'ruid'} (i.e. Reproducible Unique IDs), the function concatenates
location, year, trial name, plot, row and column IDs. if \code{get_unique_id = 'uuid'} 
(i.e. Universal Unique IDs), the function generates UUIDs by time randomly. 
Note that UUIDs are uninformative and not reproducible.

If input fieldbook already has unique IDs for each plot, the \code{get_unique_id} 
argument should be set to 'custom'; and the \code{unique_id}
argument must be specified as a string using the column name in the input fieldbook
that denotes plot unique IDs.

if \code{Year} is NULL, the function uses the current year as defined in the sys.time().

If the user is printing labels for any incomplete block design, the imported
fieldbook must include an IBLOCK column if the users wishes to display intra-blocking
information for experimental plots on the label.

Set the \code{IBlock} argument to TRUE if the field layout has incomplete
blocks within replications. The imported  fieldbook must include an IBLOCK column if
the \code{IBlock} argument is set to TRUE.
}
\examples{

\dontrun{
Generate simple field plot labels using the square_lattice sample fieldbook
library(qrlabelr)
field_label(
dat = square_lattice,
font_sz = 10,
IBlock = FALSE,
get_unique_id = "ruid",
rname = "Adoma",
family = "sans",
seed_source = FALSE)
}

\dontrun{
Generate treetag labels using the square_lattice sample fieldbook
field_label(
dat = square_lattice,
wdt = 1,
hgt = 6,
top_mar = 0,
bot_mar = 0,
left_mar = 0,
right_mar = 0,
page_hgt = 6,
page_wdt = 4,
numrow = 1,
numcol = 4,
get_unique_id = "ruid",
font_sz = 10,
IBlock = TRUE,
rounded = FALSE,
rname = "AWK",
family = "sans",
Treetag = TRUE,
seed_source = TRUE,
seed_source_id = "SEED_SOURCE")
}

}
\seealso{
\code{\link{create_label}} and \code{\link{gp_label}}
}
