% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_new_functions.R
\name{gp_label}
\alias{gp_label}
\title{Create a general-purpose (gp) label}
\usage{
gp_label(
  dat,
  get_unique_id = c("uuid", "custom"),
  unique_id = NULL,
  top_left_txt1 = NULL,
  top_left_txt2 = NULL,
  top_right_txt1 = NULL,
  top_right_txt2 = NULL,
  center_right_txt1 = NULL,
  center_right_txt2 = NULL,
  center_right_txt3 = NULL,
  bottom_left_txt1 = NULL,
  bottom_left_txt2 = NULL,
  top_left_id1 = NULL,
  top_left_id2 = NULL,
  top_right_id1 = NULL,
  top_right_id2 = NULL,
  center_right_id1 = NULL,
  center_right_id2 = NULL,
  center_right_id3 = NULL,
  bottom_left_id1 = NULL,
  bottom_left_id2 = NULL,
  ...
)
}
\arguments{
\item{dat}{is an input data frame of fieldbook that contains plot attributes.}

\item{get_unique_id}{is to be set to 'uuid' if universal unique ids are to be generated.}

\item{unique_id}{is a vector containing unique identifiers or strings to generate QR codes.
from imported fieldbook. Set to 'custom' if imported fieldbook already has unique IDs for each plot.}

\item{top_left_txt1}{is the prefix text for top-left row 1.}

\item{top_left_txt2}{is the prefix text for top-left row 2.}

\item{top_right_txt1}{is the prefix text for the top-right row 1.}

\item{top_right_txt2}{is the prefix text for the top-right row 2.}

\item{center_right_txt1}{is the prefix text for center-right row 1.}

\item{center_right_txt2}{is the prefix text for center-right row 2.}

\item{center_right_txt3}{is the prefix text for center-right row 3.}

\item{bottom_left_txt1}{is the column identifier in dat containing text for bottom-left row 1.}

\item{bottom_left_txt2}{is the column identifier in dat containing text for bottom-left row 2.}

\item{top_left_id1}{is the column identifier in dat containing text for top-left row 1.}

\item{top_left_id2}{is the column identifier in dat containing text for top-left row 2.}

\item{top_right_id1}{is the column identifier in dat containing text for top-right row 1.}

\item{top_right_id2}{is the column identifier in dat containing text for top-right row 2.}

\item{center_right_id1}{is the column identifier in dat containing text for center-right row 1.}

\item{center_right_id2}{is the column identifier in dat containing text for center-right row 2.}

\item{center_right_id3}{is the column identifier in dat containing text for center-right row 3.}

\item{bottom_left_id1}{is the column identifier in dat containing text for bottom-left row 1.}

\item{bottom_left_id2}{is the column identifier in dat containing text for bottom-left row 2.}

\item{...}{additional arguments passed to the \code{create_label} function.}
}
\value{
A PDF file containing plot labels affixed with QR codes, and
 an updated fieldbook-- all saved to the default or set working directory.
}
\description{
This function gives more flexibility to the user to design any general-purpose
label affixed with QR codes. It gives a lot of control to the user with 
respect to what human-readable text items gets displayed on the label. 
Arguments are passed to the `create_label()` function.
}
\examples{

\dontrun{
Generate general-purpose label using the square_lattice sample fieldbook.
library(qrlabelr)
gp_label(dat = square_lattice,
get_unique_id = "uuid",
font_sz = 10,
family = "sans",
top_left_txt1 = 'Plot:',
top_left_txt2 = 'Row:',
top_right_txt1 = 'Rep:',
top_right_txt2 = 'Col:',
center_right_txt1 = 'iBlock:',
center_right_txt2 = 'Seed:',
center_right_txt3 = 'AWK',
bottom_left_txt1 = 'Loc:',
top_left_id1 = 'PLOT',
top_left_id2 = 'ROW',
top_right_id1 = 'REP',
top_right_id2 = 'COLUMN',
center_right_id1 = 'IBLOCK',
center_right_id2 = 'SEED_SOURCE',
bottom_left_id1 = 'LOCATION',
bottom_left_id2 = 'TREATMENT'
)
}

}
\seealso{
\code{\link{create_label}} and \code{\link{field_label}}
}
